test(messaging): split topology and roster coverage
